move GitbaseSessionBuilder to the root package
